# Process performance improvements