--- a/src/Data/Machine/Source.hs
+++ b/src/Data/Machine/Source.hs
@@ -31,7 +31,7 @@ import Data.Foldable
 import Data.Machine.Plan
 import Data.Machine.Type
 import Data.Machine.Process
-import Prelude (Enum, Eq, Int, Maybe, Monad, ($))
+import Prelude (Enum, Int, Maybe, Monad, ($))
 
 -------------------------------------------------------------------------------
 -- Source
@@ -128,7 +128,7 @@ replicated n x = repeated x ~> taking n
 -- >>> run $ enumerateFromTo 1 3
 -- [1,2,3]
 --
-enumerateFromTo :: (Enum a, Eq a) => a -> a -> Source a
+enumerateFromTo :: Enum a => a -> a -> Source a
 enumerateFromTo start end = source [ start .. end ]
 
 -- | 'unfold' @k seed@ The function takes the element and returns Nothing if it