… benchmarks/Benchmarks.hs
… examples/Examples.hs
… src/Data/Machine/Stack.hs
… src/Data/Machine/Tee.hs
… src/Data/Machine/Type.hs
… src/Data/Machine/Wye.hs