This one might be a little controversial, but I did some experiments rewriting `Process` and `Source` `Machines` explicitly instead of using a `Plan` and got performance improvements essentially across the board.
I also believe there were some bugs in `buffered`, as there were situations where it would lead to empty list outputs.

This does make these machines less readable and harder to fathom however.

Here's a handful of before and afters

```
Before:
benchmarking map/machines
time                 82.80 ms   (72.96 ms .. 93.38 ms)
                     0.977 R²   (0.954 R² .. 0.996 R²)
mean                 83.58 ms   (79.74 ms .. 88.00 ms)
std dev              6.704 ms   (5.518 ms .. 8.134 ms)
variance introduced by outliers: 19% (moderately inflated)

After:
benchmarking map/machines
time                 74.91 ms   (68.96 ms .. 78.67 ms)
                     0.990 R²   (0.971 R² .. 0.999 R²)
mean                 74.33 ms   (72.12 ms .. 76.15 ms)
std dev              3.490 ms   (2.413 ms .. 4.940 ms)

Before:
benchmarking dropWhile/machines
time                 47.15 ms   (45.25 ms .. 49.08 ms)
                     0.995 R²   (0.990 R² .. 0.999 R²)
mean                 50.47 ms   (48.74 ms .. 55.15 ms)
std dev              5.309 ms   (1.213 ms .. 9.059 ms)

After:
benchmarking dropWhile/machines
time                 36.67 ms   (35.80 ms .. 37.54 ms)
                     0.998 R²   (0.997 R² .. 1.000 R²)
mean                 36.58 ms   (36.11 ms .. 37.13 ms)
std dev              1.001 ms   (761.2 μs .. 1.369 ms)

Before:
benchmarking scan/machines
time                 120.0 ms   (108.2 ms .. 131.1 ms)
                     0.989 R²   (0.974 R² .. 1.000 R²)
mean                 115.0 ms   (112.8 ms .. 120.0 ms)
std dev              4.770 ms   (1.391 ms .. 7.354 ms)

After:
benchmarking scan/machines
time                 81.67 ms   (78.75 ms .. 87.71 ms)
                     0.993 R²   (0.980 R² .. 0.999 R²)
mean                 82.60 ms   (80.38 ms .. 84.80 ms)
std dev              3.406 ms   (2.520 ms .. 4.855 ms)

Before:
benchmarking takeWhile/machines
time                 116.7 ms   (113.7 ms .. 119.8 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 109.8 ms   (106.7 ms .. 112.4 ms)
std dev              4.433 ms   (2.954 ms .. 6.029 ms)

After:
benchmarking takeWhile/machines
time                 75.94 ms   (71.89 ms .. 79.91 ms)
                     0.995 R²   (0.984 R² .. 0.999 R²)
mean                 78.85 ms   (77.01 ms .. 80.78 ms)
std dev              3.132 ms   (2.149 ms .. 4.631 ms)

Before:
benchmarking mapM/machines
time                 125.3 ms   (120.7 ms .. 131.9 ms)
                     0.998 R²   (0.995 R² .. 1.000 R²)
mean                 114.0 ms   (109.4 ms .. 117.9 ms)
std dev              6.277 ms   (4.816 ms .. 7.943 ms)

After:
benchmarking mapM/machines
time                 83.82 ms   (80.62 ms .. 86.54 ms)
                     0.997 R²   (0.992 R² .. 0.999 R²)
mean                 84.40 ms   (82.85 ms .. 86.16 ms)
std dev              2.700 ms   (2.061 ms .. 3.571 ms)
```